requirements.txt
aiogram>=3.4,<4.0
python-dotenv>=1.0
httpx>=0.27

fastapi>=0.115
uvicorn[standard]>=0.30
